014-contract-risk-agent/pyproject.toml
[project]
name = "contract-risk-agent"
version = "0.1.0"
description = "Simple multi-contract review app"
readme = "README.md"
authors = [
    { name = "rishisan99", email = "santhoshrishi9999@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "pydantic>=2.13.3",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.27",
    "requests>=2.33.1",
    "streamlit>=1.57.0",
    "uvicorn>=0.46.0",
]

[project.scripts]
contract-risk-agent = "contract_risk_agent:main"

[build-system]
requires = ["uv_build>=0.11.8,<0.12.0"]
build-backend = "uv_build"
